third_party/blink/web_tests/external/wpt/editing/run/insertorderedlist_1-1000-expected.txt

This is a testharness.js-based test.
Found 235 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "foo[bar]baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "foo[bar]baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "foo<br>[bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<ol><li>bar</li></ol>" but got "foo<br><ol><li>bar</li></ol>"
[FAIL] [["insertorderedlist",""]] "foo<br>[bar]" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "foo<br>[bar]" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "f[oo<br>b]ar<br>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "f[oo<br>b]ar<br>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<p>[foo]<br>bar</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>bar</p>" but got "<p><ol><li>foo</li></ol>bar</p>"
[FAIL] [["insertorderedlist",""]] "<p>[foo]<br>bar</p>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<p>[foo]<br>bar</p>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "[foo<ol><li>bar]</ol>baz" queryCommandIndeterm("insertorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertorderedlist",""]] "[foo<ol><li>bar]</ol>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "[foo<ol><li>bar]</ol>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "foo<ol><li>[bar</ol>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<ol><li>bar</li><li>baz</li></ol>" but got "foo<br>bar<br><ol><li>baz</li></ol>"
[FAIL] [["insertorderedlist",""]] "foo<ol><li>[bar</ol>baz]" queryCommandIndeterm("insertorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertorderedlist",""]] "foo<ol><li>[bar</ol>baz]" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "foo<ol><li>[bar</ol>baz]" queryCommandState("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertorderedlist",""]] "foo<ol><li>[bar</ol>baz]" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "[foo<ul><li>bar]</ul>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "[foo<ul><li>bar]</ul>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "foo<ul><li>[bar</ul>baz]" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "foo<ul><li>[bar</ul>baz]" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "foo<ul><li>[bar</ul><ol><li>baz]</ol>quz" queryCommandIndeterm("insertorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertorderedlist",""]] "foo<ul><li>[bar</ul><ol><li>baz]</ol>quz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "foo<ul><li>[bar</ul><ol><li>baz]</ol>quz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "foo<ol><li>[bar</ol><ul><li>baz]</ul>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<ol><li>bar</li><li>baz</li></ol>quz" but got "foo<br>bar<br><ol><li>baz</li></ol>quz"
[FAIL] [["insertorderedlist",""]] "foo<ol><li>[bar</ol><ul><li>baz]</ul>quz" queryCommandIndeterm("insertorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertorderedlist",""]] "foo<ol><li>[bar</ol><ul><li>baz]</ul>quz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "foo<ol><li>[bar</ol><ul><li>baz]</ul>quz" queryCommandState("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertorderedlist",""]] "foo<ol><li>[bar</ol><ul><li>baz]</ul>quz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<table><tbody><tr><td>fo[o<td>b]ar<td>baz</table>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<table><tbody><tr><td>fo[o<td>b]ar<td>baz</table>" queryCommandState("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertorderedlist",""]] "<table><tbody><tr><td>fo[o<td>b]ar<td>baz</table>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>foo<p>[bar]<p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ol><li>bar</li></ol><p>baz</p>" but got "<p>foo</p><p><ol><li>bar</li></ol></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>foo<p>[bar]<p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>foo<p>[bar]<p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>foo<p>[bar]<p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ol><li>bar</li></ol><p>baz</p>" but got "<p>foo</p><p><ol><li>bar</li></ol></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>foo<p>[bar]<p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>foo<p>[bar]<p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<p>foo<blockquote>[bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<p>foo<blockquote>[bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<dl><dt>foo<dd>[bar]<dt>baz<dd>quz</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd><ol><li>bar</li></ol></dd><dt>baz</dt><dd>quz</dd></dl>" but got "<dl><dt>foo</dt><ol><li>bar</li></ol><dt>baz</dt><dd>quz</dd></dl>"
[FAIL] [["insertorderedlist",""]] "<dl><dt>foo<dd>[bar]<dt>baz<dd>quz</dl>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<dl><dt>foo<dd>[bar]<dt>baz<dd>quz</dl>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<dl><dt>foo<dd>bar<dt>[baz]<dd>quz</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd>bar</dd><dt><ol><li>baz</li></ol></dt><dd>quz</dd></dl>" but got "<dl><dt>foo</dt><dd>bar</dd><ol><li>baz</li></ol><dd>quz</dd></dl>"
[FAIL] [["insertorderedlist",""]] "<dl><dt>foo<dd>bar<dt>[baz]<dd>quz</dl>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<dl><dt>foo<dd>bar<dt>[baz]<dd>quz</dl>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo<p>bar]<p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>bar</li></ol><p>baz</p>" but got "<p><ol><li>foo</li><li>bar</li></ol></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo<p>bar]<p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo<p>bar]<p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo<p>bar]<p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>bar</li></ol><p>baz</p>" but got "<p><ol><li>foo</li><li>bar</li></ol></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo<p>bar]<p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo<p>bar]<p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo<blockquote>bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><p>baz</p>" but got "<p><ol><li>foo</li><li>bar</li></ol></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo<blockquote>bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo<blockquote>bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo<blockquote>bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><p>baz</p>" but got "<p><ol><li>foo</li><li>bar</li></ol></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo<blockquote>bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo<blockquote>bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<dl><dt>[foo<dd>bar]<dt>baz<dd>quz</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt><ol><li>foo</li></ol></dt><dd><ol><li>bar</li></ol></dd><dt>baz</dt><dd>quz</dd></dl>" but got "<dl><ol><li>foo</li><dd>bar</dd></ol><dt>baz</dt><dd>quz</dd></dl>"
[FAIL] [["insertorderedlist",""]] "<dl><dt>[foo<dd>bar]<dt>baz<dd>quz</dl>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<dl><dt>[foo<dd>bar]<dt>baz<dd>quz</dl>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<dl><dt>foo<dd>[bar<dt>baz]<dd>quz</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd><ol><li>bar</li></ol></dd><dt><ol><li>baz</li></ol></dt><dd>quz</dd></dl>" but got "<dl><dt>foo</dt><ol><li>bar</li><dt>baz</dt></ol><dd>quz</dd></dl>"
[FAIL] [["insertorderedlist",""]] "<dl><dt>foo<dd>[bar<dt>baz]<dd>quz</dl>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<dl><dt>foo<dd>[bar<dt>baz]<dd>quz</dl>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo<blockquote><p>bar]<p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><blockquote><p>baz</p></blockquote>" but got "<p><ol><li>foo</li><li>bar</li></ol></p><blockquote><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo<blockquote><p>bar]<p>baz</blockquote>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo<blockquote><p>bar]<p>baz</blockquote>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo<blockquote><p>bar]<p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><blockquote><p>baz</p></blockquote>" but got "<p><ol><li>foo</li><li>bar</li></ol></p><blockquote><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo<blockquote><p>bar]<p>baz</blockquote>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo<blockquote><p>bar]<p>baz</blockquote>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><div>bar</div><ol><li>baz</li></ol>" but got "<ol><li>foo</li></ol>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]<li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]<li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>bar</p><ol><li>baz</li></ol>" but got "<ol><li>foo</li></ol>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]<li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]<li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</ol>[bar]" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</ol>[bar]" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "[foo]<ol><li>bar</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "[foo]<ol><li>bar</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</ol>[bar]<ol><li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</ol>[bar]<ol><li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><ol><li>[foo]</ol></ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><ol><li>[foo]</ol></ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo]<br>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<br>bar</div><ol><li>baz</li></ol>" but got "foo<br>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo]<br>bar<li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo]<br>bar<li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo]<br>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<br>bar</p><ol><li>baz</li></ol>" but got "foo<br>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo]<br>bar<li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo]<br>bar<li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<br>[bar]<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<br>bar</div><ol><li>baz</li></ol>" but got "foo<br>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<br>[bar]<li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<br>[bar]<li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<br>[bar]<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<br>bar</p><ol><li>baz</li></ol>" but got "foo<br>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<br>[bar]<li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<br>[bar]<li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li><div>[foo]</div>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div>bar</div><ol><li>baz</li></ol>" but got "<div style=\\"display:inline !important\\">foo</div>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li><div>[foo]</div>bar<li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li><div>[foo]</div>bar<li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li><div>[foo]</div>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><p>bar</p><ol><li>baz</li></ol>" but got "<div style=\\"display:inline !important\\">foo</div>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li><div>[foo]</div>bar<li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li><div>[foo]</div>bar<li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo<ol><li>[bar]<li>baz</ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>bar</li><ol><li>baz</li></ol><li>quz</li></ol>" but got "<ol><li>foo<li>bar</li><ol><li>baz</li></ol></li><li>quz</li></ol>"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo<ol><li>[bar]<li>baz</ol><li>quz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo<ol><li>[bar]<li>baz</ol><li>quz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo<ol><li>bar<li>[baz]</ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol><li>baz</li><li>quz</li></ol>" but got "<ol><li>foo<ol><li>bar</li></ol><li>baz</li></li><li>quz</li></ol>"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo<ol><li>bar<li>[baz]</ol><li>quz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo<ol><li>bar<li>[baz]</ol><li>quz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</li><ol><li>[bar]<li>baz</ol><li>quz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</li><ol><li>[bar]<li>baz</ol><li>quz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</li><ol><li>bar<li>[baz]</ol><li>quz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</li><ol><li>bar<li>[baz]</ol><li>quz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ol><ol><li>bar</li></ol><li>baz</li></ol>" but got "foo<ol><li>bar</li></ol><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol><li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol><li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ol><ol><li>bar</li></ol><li>baz</li></ol>" but got "foo<ol><li>bar</li></ol><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol><li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol><li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo]</li><ol><li>bar</ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ol><ol><li>bar</li></ol><li>baz</li></ol>" but got "foo<br><ol><ol><li>bar</li></ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo]</li><ol><li>bar</ol><li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo]</li><ol><li>bar</ol><li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo]</li><ol><li>bar</ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ol><ol><li>bar</li></ol><li>baz</li></ol>" but got "foo<br><ol><ol><li>bar</li></ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo]</li><ol><li>bar</ol><li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo]</li><ol><li>bar</ol><li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]<ol><li>baz</ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><div>bar</div><ol><ol><li>baz</li></ol><li>quz</li></ol>" but got "<ol><li>foo</li></ol>bar<ol><li>baz</li></ol><ol><li>quz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]<ol><li>baz</ol><li>quz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]<ol><li>baz</ol><li>quz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]<ol><li>baz</ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>bar</p><ol><ol><li>baz</li></ol><li>quz</li></ol>" but got "<ol><li>foo</li></ol>bar<ol><li>baz</li></ol><ol><li>quz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]<ol><li>baz</ol><li>quz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]<ol><li>baz</ol><li>quz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]</li><ol><li>baz</ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><div>bar</div><ol><ol><li>baz</li></ol><li>quz</li></ol>" but got "<ol><li>foo</li></ol>bar<br><ol><ol><li>baz</li></ol><li>quz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]</li><ol><li>baz</ol><li>quz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]</li><ol><li>baz</ol><li>quz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]</li><ol><li>baz</ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>bar</p><ol><ol><li>baz</li></ol><li>quz</li></ol>" but got "<ol><li>foo</li></ol>bar<br><ol><ol><li>baz</li></ol><li>quz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]</li><ol><li>baz</ol><li>quz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<li>[bar]</li><ol><li>baz</ol><li>quz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<ol><li>bar<li>baz</ol><li>[quz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<ol><li>bar</li><li>baz</li></ol></li></ol><div>quz</div>" but got "<ol><li>foo<ol><li>bar</li><li>baz</li></ol></li></ol>quz"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<ol><li>bar<li>baz</ol><li>[quz]</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<ol><li>bar<li>baz</ol><li>[quz]</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<ol><li>bar<li>baz</ol><li>[quz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<ol><li>bar</li><li>baz</li></ol></li></ol><p>quz</p>" but got "<ol><li>foo<ol><li>bar</li><li>baz</li></ol></li></ol>quz"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<ol><li>bar<li>baz</ol><li>[quz]</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<ol><li>bar<li>baz</ol><li>[quz]</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo</li><ol><li>bar<li>baz</ol><li>[quz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li><li>baz</li></ol></ol><div>quz</div>" but got "<ol><li>foo</li><ol><li>bar</li><li>baz</li></ol></ol>quz"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo</li><ol><li>bar<li>baz</ol><li>[quz]</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo</li><ol><li>bar<li>baz</ol><li>[quz]</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo</li><ol><li>bar<li>baz</ol><li>[quz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li><li>baz</li></ol></ol><p>quz</p>" but got "<ol><li>foo</li><ol><li>bar</li><li>baz</li></ol></ol>quz"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo</li><ol><li>bar<li>baz</ol><li>[quz]</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo</li><ol><li>bar<li>baz</ol><li>[quz]</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<li>[bar<li>baz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><div>bar</div><div>baz</div>" but got "<ol><li>foo</li></ol>bar<br>baz"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<li>[bar<li>baz]</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<li>[bar<li>baz]</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<li>[bar<li>baz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>bar</p><p>baz</p>" but got "<ol><li>foo</li></ol>bar<br>baz"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<li>[bar<li>baz]</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<li>[bar<li>baz]</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ol><li>bar</li><li>baz</li></ol>" but got "foo<br>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" queryCommandState("insertorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" queryCommandIndeterm("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ol><li>bar</li><li>baz</li></ol>" but got "foo<br>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" queryCommandState("insertorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" queryCommandIndeterm("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>bar</li></ol><div>baz</div>" but got "<ol><li>foo<li>bar</li></li></ol>baz"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" queryCommandState("insertorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" queryCommandIndeterm("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>bar</li></ol><p>baz</p>" but got "<ol><li>foo<li>bar</li></li></ol>baz"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" queryCommandState("insertorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" queryCommandIndeterm("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar</ol><li>baz]</ol><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ol><li>bar</li></ol><div>baz</div><p>extra</p>" but got "foo<br>bar<br>baz<br><p>extra</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar</ol><li>baz]</ol><p>extra" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar</ol><li>baz]</ol><p>extra" queryCommandIndeterm("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar</ol><li>baz]</ol><p>extra" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar</ol><li>baz]</ol><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ol><li>bar</li></ol><p>baz</p><p>extra</p>" but got "foo<br>bar<br>baz<br><p>extra</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar</ol><li>baz]</ol><p>extra" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar</ol><li>baz]</ol><p>extra" queryCommandIndeterm("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar</ol><li>baz]</ol><p>extra" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ol><ol><li>bar</li></ol><li>baz</li></ol>" but got "foo<ol><li>bar</li></ol>baz"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ol><ol><li>bar</li></ol><li>baz</li></ol>" but got "foo<ol><li>bar</li></ol>baz"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo<ol><li>[bar]</ol>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>bar</li><li>baz</li></ol>" but got "<ol><li>foo<li>bar</li>baz</li></ol>"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo<ol><li>[bar]</ol>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo<ol><li>[bar]</ol>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<ol><li>bar</ol>[baz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><div>baz</div>" but got "foo<ol><li>bar</li></ol>baz"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<ol><li>bar</ol>[baz]</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo<ol><li>bar</ol>[baz]</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<ol><li>bar</ol>[baz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><p>baz</p>" but got "foo<ol><li>bar</li></ol>baz"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<ol><li>bar</ol>[baz]</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo<ol><li>bar</ol>[baz]</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ol><li>bar</li><li>baz</li></ol>" but got "foo<br>bar<br>baz"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" queryCommandState("insertorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" queryCommandIndeterm("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ol><li>bar</li><li>baz</li></ol>" but got "foo<br>bar<br>baz"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" queryCommandState("insertorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" queryCommandIndeterm("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo</ul>[bar]" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo</ul>[bar]" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "[foo]<ul><li>bar</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "[foo]<ul><li>bar</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo</ul>[bar]<ul><li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo</ul>[bar]<ul><li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><ul><li>[foo]</ul></ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li>foo</li></ol></ol>" but got "<ul><ol><li>foo</li></ol></ul>"
[FAIL] [["insertorderedlist",""]] "<ul><ul><li>[foo]</ul></ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><ul><li>[foo]</ul></ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo]<br>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<br>bar</li></ol><ul><li>baz</li></ul>" but got "<ol><li>foo</li></ol>bar<br><ul><li>baz</li></ul>"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo]<br>bar<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo]<br>bar<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<br>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<br>bar</li></ol><ul><li>baz</li></ul>" but got "foo<br><ol><li>bar</li></ol><ul><li>baz</li></ul>"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<br>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<br>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li><div>[foo]</div>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><div>foo</div>bar</li></ol><ul><li>baz</li></ul>" but got "<ol><li><div style=\\"display:inline !important\\">foo</div>bar</li></ol><ul><li>baz</li></ul>"
[FAIL] [["insertorderedlist",""]] "<ul><li><div>[foo]</div>bar<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li><div>[foo]</div>bar<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>[bar]<li>baz</ul><li>quz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ol><li>bar</li></ol><ul><li>baz</li></ul><li>quz</li></ul>" but got "<ul><li>foo<ol><li>bar</li></ol><ul><li>baz</li></ul></li><li>quz</li></ul>"
Harness: the test ran to completion.
